[WIP] Create a lexer and parser for Dellos10
Dellos10 has syntax that is somewhat similar to Cisco, so when feeding
in configurations to batfish some of the configuration gets parsed
correctly, but other parts do not.

So, instead of extending the Cisco lexer and parser like Arista did (and
subsequently moved into its own files) let's go ahead and start a new
lexer and parser file for Dellos10.